generalize gpu backend
