+++ mlx/backend/no_gpu/fence.cpp
