+++ mlx/device.cpp
@@ -1,13 +1,15 @@
 // Copyright © 2023 Apple Inc.
 
+#include <stdexcept>
+
+#include "mlx/backend/cpu/available.h"
+#include "mlx/backend/gpu/available.h"
 #include "mlx/device.h"
-#include "mlx/backend/metal/metal.h"
 
 namespace mlx::core {
 
 Device& mutable_default_device() {
-  static Device default_device{
-      metal::is_available() ? Device::gpu : Device::cpu};
+  static Device default_device{gpu::is_available() ? Device::gpu : Device::cpu};
   return default_device;
 }
 
@@ -16,7 +18,7 @@ const Device& default_device() {
 }
 
 void set_default_device(const Device& d) {
-  if (!metal::is_available() && d == Device::gpu) {
+  if (!gpu::is_available() && d == Device::gpu) {
     throw std::invalid_argument(
         "[set_default_device] Cannot set gpu device without gpu backend.");
   }
@@ -29,6 +31,15 @@ bool operator==(const Device& lhs, const Device& rhs) {
 
 bool operator!=(const Device& lhs, const Device& rhs) {
   return !(lhs == rhs);
+}
+
+bool is_available(const Device& d) {
+  switch (d.type) {
+    case Device::cpu:
+      return cpu::is_available();
+    case Device::gpu:
+      return gpu::is_available();
+  }
 }
 
 } // namespace mlx::core
